+++ Coimbra/Components/LateUpdateListener.cs
@@ -1,5 +1,4 @@
-﻿using System.Runtime.CompilerServices;
-using UnityEngine;
+﻿using UnityEngine;
 
 namespace Coimbra
 {
@@ -8,11 +7,32 @@ namespace Coimbra
     /// </summary>
     [DisallowMultipleComponent]
     [AddComponentMenu(CoimbraUtility.GeneralMenuPath + "Late Update Listener")]
+    [HelpURL("https://docs.unity3d.com/ScriptReference/MonoBehaviour.LateUpdate.html")]
     public sealed class LateUpdateListener : PlayerLoopListenerBase
     {
+        /// <inheritdoc />
+        public override event EventHandler OnTrigger
+        {
+            add
+            {
+                base.OnTrigger += value;
+                enabled = HasListener;
+            }
+            remove
+            {
+                base.OnTrigger -= value;
+                enabled = HasListener;
+            }
+        }
+
+        private void Awake()
+        {
+            enabled = HasListener;
+        }
+
         private void LateUpdate()
         {
-            Trigger();
+            Trigger(Time.deltaTime);
         }
     }
 }